+++ CHANGELOG.md
@@ -14,7 +14,9 @@ The changelog format is inspired by [keep-a-changelog](https://github.com/olivie
 
 ## [Unreleased]
 - "sendto" now looks for a config file in home directory
+- "sendto" options to skip large VASP files
 - New CPLAP plotter
+
 
 ## [1.0] - 2016-04-25